Modularise issue, pullRequest events
